refactor: move SQLiteDiaclect_libsql to its own file
